Updates install scripts and READMEs

+++ deploy/vagrant/make_boxes/dtaas/provision.sh
@@ -12,18 +12,25 @@ apt-get install -y \
     gnupg \
     lsb-release \
     zsh \
-    apache2-utils
+    apache2-utils \
+    net-tools
 
 mkdir -p /etc/apt/keyrings
-curl -fsSL "https://download.docker.com/linux/ubuntu/gpg" | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-printf \
-  "deb [arch=%s signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
-  %s stable \n" "$(dpkg --print-architecture)" "$(lsb_release -cs)"  | tee /etc/apt/sources.list.d/docker.list > /dev/null
+if [ ! -f /etc/apt/keyrings/docker.gpg ]
+then
+  curl -fsSL "https://download.docker.com/linux/ubuntu/gpg" | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+  printf \
+    "deb [arch=%s signed-by=/etc/apt/keyrings/docker.gpg] \
+    https://download.docker.com/linux/ubuntu %s stable" \
+    "$(dpkg --print-architecture)" "$(lsb_release -cs)"  | \
+    tee /etc/apt/sources.list.d/docker.list > /dev/null
+fi
+
 apt-get update -y
 apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
-groupadd docker
-usermod -aG docker vagrant
-newgrp docker
+groupadd docker || true
+usermod -aG docker vagrant || true
+newgrp docker || true
 service docker start
 docker run hello-world
 
@@ -39,8 +46,11 @@ chmod 755 /usr/local/bin/docker-compose /usr/bin/docker-compose
 apt-get update
 apt-get install -y ca-certificates curl gnupg
 mkdir -p /etc/apt/keyrings
-curl -fsSL "https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key" | \
-  gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
+if [ ! -f /etc/apt/keyrings/nodesource.gpg ]
+then
+  curl -fsSL "https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key" | \
+    gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
+fi
 NODE_MAJOR=18
 printf "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | \
   tee /etc/apt/sources.list.d/nodesource.list
@@ -48,13 +58,22 @@ apt-get update
 apt-get install -y nodejs
 npm install -g npm@10.2.0
 
-curl -sL "https://dl.yarnpkg.com/debian/pubkey.gpg" | gpg --dearmor | \
-  tee /usr/share/keyrings/yarnkey.gpg >/dev/null
-printf "deb [signed-by=/usr/share/keyrings/yarnkey.gpg] https://dl.yarnpkg.com/debian stable main \n" | \
-  tee /etc/apt/sources.list.d/yarn.list
+if [ ! -f /usr/share/keyrings/yarnkey.gpg ]
+then
+  curl -sL "https://dl.yarnpkg.com/debian/pubkey.gpg" | gpg --dearmor | \
+    tee /usr/share/keyrings/yarnkey.gpg >/dev/null
+  printf "deb [signed-by=/usr/share/keyrings/yarnkey.gpg] https://dl.yarnpkg.com/debian stable main \n" | \
+    tee /etc/apt/sources.list.d/yarn.list
+fi
 apt-get update -y
 apt-get install -y yarn
 npm install -g serve
+
+# Install openssl for certificate generation
+apt-get install -y wget openssl
+
+# Install playwright tool for integration tests on browsers
+npx playwright install-deps
 
 #-------------
 printf "\n\n Install jupyterlab and mkdocs"
@@ -86,3 +105,10 @@ docker pull telegraf:1.28.2
 docker pull rabbitmq:3-management
 docker pull eclipse-mosquitto:2
 docker pull gitlab/gitlab-ce:16.4.1-ce.0
+
+# Install markdownlint
+sudo gem install mdl
+
+# Install madge for generating dependency graphs of typescript projects
+sudo apt-get install graphviz
+sudo npm install -g madge
